factor out generate12DigitId
