+++ src/utils/generate12DigitId.ts
@@ -1,0 +1,3 @@
+export function generate12DigitId() {
+  return Math.round(Math.random() * 1000000000000).toString();
+}
